+++ flaggr/src/main/java/com/comuto/flag/parsing/CustomJsonDeserializer.java
@@ -1,9 +1,11 @@
 package com.comuto.flag.parsing;
 
+import android.support.annotation.Nullable;
 import com.comuto.flag.operators.InsetOperator;
 import com.comuto.flag.operators.ModuloOperator;
 import com.comuto.flag.operators.Operator;
 import com.comuto.flag.operators.PercentageOperator;
+import com.comuto.flag.operators.UnknownOperator;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
@@ -27,8 +29,10 @@ public class CustomJsonDeserializer implements JsonDeserializer<Operator> {
     private static final String PERCENTAGE_OPERATOR_PACKAGE = PercentageOperator.class.getCanonicalName();
     private static final String MODULO_KEY = "modulo";
     private static final String MODULO_OPERATOR_PACKAGE = ModuloOperator.class.getCanonicalName();
+    private static final String UNKNOWN_OPERATOR_PACKAGE = UnknownOperator.class.getCanonicalName();
     private static final String JSON_NAME_ATTR = "name";
 
+    @Nullable
     @Override
     public Operator deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
         final JsonObject jsonObject = json.getAsJsonObject();
@@ -44,8 +48,7 @@ public class CustomJsonDeserializer implements JsonDeserializer<Operator> {
                 } else if (type.equals(MODULO_KEY)) {
                     return context.deserialize(jsonObject, Class.forName(MODULO_OPERATOR_PACKAGE));
                 }
-                throw new JsonParseException("Unknown Operator : " + type);
-
+                return context.deserialize(jsonObject, Class.forName(UNKNOWN_OPERATOR_PACKAGE));
             } catch (ClassNotFoundException cnfe) {
                 throw new JsonParseException("Unknown element type: " + type, cnfe);
             }