+++ flaggr/src/main/java/com/comuto/flag/strategy/MajorityStrategy.java
@@ -21,23 +21,23 @@ public class MajorityStrategy implements Strategy {
      * @return true if a majority of conditions are validated (> half), false otherwise
      */
     @Override
-    public boolean isFlagActivated(Flag flag, @NonNull FlagContextInterface context) {
+    public Flag.FlagResultStatus getFlagStatus(Flag flag, @NonNull FlagContextInterface context) {
         int verifiedCondition = 0;
         List<Condition> conditionList = flag.getConditions();
         if (null != conditionList) {
             for (Condition condition : conditionList) {
                 if (null != condition
                     && null != condition.getOperator()
-                    && condition.getOperator().appliesTo(context.getValue(condition.getKey()))) {
+                    && condition.getOperator().appliesTo(context.getValue(condition.getKey())) == Flag.FlagResultStatus.ENABLED) {
                     ++verifiedCondition;
                 }
             }
 
             if (verifiedCondition > (conditionList.size() / 2)) {
-                return true;
+                return Flag.FlagResultStatus.ENABLED;
             }
         }
 
-        return false;
+        return Flag.FlagResultStatus.DISABLED;
     }
 }