+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2016-12-27
+- The result of checking if a flag is activated is now FlagResult
+- This value can be ENABLED, DISABLED and UNKNOWN
+- If the downloaded Json is empty, it will not override the current configuration of flag
 
 ## [0.0.3] - 2016-05-26
 ### Added